Small refactor to make CompoundStatement an inner class.
